Moved started msg to manager
